requirements.txt
tensorflow-gpu==1.15
tensorboard
absl-py
astor
backports.functools-lru-cache
backports.weakref
bleach
boto
boto3
botocore
bz2file
certifi
chardet
contextlib2
cycler
Cython
docutils
enum34
funcsigs
gast
gensim
grpcio
h5py
html5lib
idna
jmespath
Keras
Keras-Applications
Keras-Preprocessing
kiwisolver
Markdown
matplotlib
mock
numpy
pbr
Pillow
protobuf
psutil
pycocotools
pyparsing
Pypubsub
python-dateutil
pytz
PyYAML
requests
s3transfer
scikit-learn
scipy
six
sklearn
smart-open
subprocess32
termcolor
Theano
typing
urllib3
Werkzeug
XlsxWriter
opencv-python
multiprocess
thread6